Implement model validation, merge VIMObjectMapper directly into VimeoNetworking
This commit does two things:

1. Merges VIMObjectMapper into VimeoNetworking as a direct import.
This facilitated making changes that were necessary for implementing
validation.  I would have done this as a separate commit if I realized
at the beginning it would be happening, but it only became apparent
when it was too late, so here we are :T.

2. Implements model object validation.  A new method is added to all
instances of VIMModelObject: validateModel.  This gives model objects
an opportunity to throw an error if they deem themselves invalid.  As
our first steps into this process, we have `VIMVideo` checking for
non-nil `uri` and `resourceKey`, and `VIMUser` checking for non-nil
`uri`.  Tests were added for these validations as well.